+++ cmd/.snapshots/TestCreateProjectFiles-client-main.py
@@ -40,7 +40,7 @@ async def main():
         print(f"Trial '{control_session.get_trial_id()}' terminating")
         await control_session.terminate_trial()
 
-    await context.start_trial(endpoint="orchestrator:9000", impl=trial_controler, trial_config=TrialConfig())
+    await context.start_trial(endpoint=cogment.Endpoint("orchestrator:9000"), impl=trial_controler, trial_config=TrialConfig())
 
 if __name__ == '__main__':
     asyncio.run(main())