+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `cogment init` now supports the simplified event data structure in actor and environment event loops.
 - `cogment init` now uses `cogment.Endpoint` and `cogment.ServedEndpoint` instead of raw TCP ports.
 
+### Fixed
+
+- The default `cogment run start` properly attach to the actors and environment services.
+
 ## v1.0.0-alpha9 - 2021-03-01
 
 ## v1.0.0-alpha8 - 2021-02-25